requirements.txt
streamlit
watchdog
sentence-transformers
pillow
torch